Use case/esac for grouped options, name customisation
Instead of a single if..fi for each of the grouped arguments it uses a
case..esac format now. This allows for better readability and easier
modifications, when needed.

Also changed every line of the script that read 'nani' (like the --help
output) to use '${0##*/}' instead, which will use the name of the script.